requirements.txt
gTTS==2.5.3 
aiogram==3.12.0
python-dotenv==1.0.1  